Updated screen coordinate values
Changed the screen coordinate system to reflect the window size and always return (0, 0) as the center position (position of player). This meant a need to update the values of the animation tree. Overall, brings it closer in line with the movement system (negative values mean left and up directions relative to center).