Additional arrows + spread
Added the ability for additional arrows to be configurable in the event the player gets a powerup that allows them to do a multi-shot. Each additional arrow adds additional spread to the shot, with the first arrow being accurate at full draw, and each additional arrow having a deviance of up to 0.1 radians per arrow.