Folders restructured
Started to define a more unified and common sense approach to folder structure. Haven't done all folders yet (namely haven't touched the Wizard folder). Inside folders you should find only the scenes relevant to that directory and other directories used as assets to those scenes. For example, inside the 'Enemies' directory is a folder called 'Bat'. Inside that folder is the bat's scene, as well as other folders for its assets such as the 'Attack' files, 'Scripts' files, and 'Sprite' files.

Godot likely will try to create scripts in the same folder as the scene, but manually change it to this setup as it will make going through the files much simpler.