Updated bat texture and readded fireballs
Added the ability for a super bat to spawn instead of a regular bat. These bats are red instead of purple and have more HP, more speed, and can fire fireballs at the player. Removed fireball clipping with walls as this is an issue if the bat has a wall directly above it.

Also added a timer delay to the fireballs so that if there is more than 1 bat then they will shoot their fireballs at different times.